+++ packages/manager/apps/telecom/src/app/telecom/pack/dashboard/pack-main.view.html
