+++ packages/manager/apps/telecom/src/app/telecom/pack/dashboard/translations/Messages_de_DE.json
