+++ packages/manager/apps/telecom/src/app/telecom/pack/dashboard/translations/Messages_en_GB.json
@@ -10,27 +10,27 @@
   "pack_xdsl_avalaibleSlot_0": "Vous avez consommé tout vos {{packType}} sur ce pack",
   "pack_xdsl_avalaibleSlot_1": "{{avalaible}} de disponible",
   "pack_xdsl_avalaibleSlot_n": "{{avalaible}} de disponibles",
-  "pack_xdsl_access_services_0": "No hay accesos",
-  "pack_xdsl_access_description": "Descripción",
-  "pack_xdsl_access_type": "Tipo",
-  "pack_xdsl_access_accessName": "Identifiant",
-  "pack_xdsl_voipLine_domain": "Dominio",
-  "pack_xdsl_voipLine_billingAccount": "Compte",
-  "pack_xdsl_hubic_domain": "Dominio",
-  "pack_xdsl_hubic_size": "Taille",
+  "pack_xdsl_access_services_0": "No access",
+  "pack_xdsl_access_description": "Description",
+  "pack_xdsl_access_type": "Type",
+  "pack_xdsl_access_accessName": "ID",
+  "pack_xdsl_voipLine_domain": "Domain",
+  "pack_xdsl_voipLine_billingAccount": "Account",
+  "pack_xdsl_hubic_domain": "Domain",
+  "pack_xdsl_hubic_size": "Size",
   "pack_xdsl_hubic_voucher": "(Code : {{voucher}})",
   "pack_xdsl_hubic_is_used": "Déja utilisé",
   "pack_xdsl_hubic_used_voucher_pending_details": "Récupération des données...",
   "pack_xdsl_hubic_used_email": "Adresse email : {{email}}",
   "pack_xdsl_hubic_used_email_unavailable": "Donnée indisponible",
   "pack_xdsl_hubic_domain_details_error": "Une erreur est survenue lors de la récupération du détail des domaines",
-  "xdslAccess": "Accès Internet",
+  "xdslAccess": "Internet access ",
   "xdslAccess_default_sidebar_title": "Mes accès",
-  "voipLine": "Téléphonie",
-  "domain": "Nom de domaine",
+  "voipLine": "Telephony ",
+  "domain": "Domain name",
   "voipEcoFax": "EcoFax Pro",
   "hubic": "HubiC",
-  "task": "Opérations",
+  "task": "Operations",
   "pack_xdsl_hubic_activate": "Profiter de mon offre hubiC",
   "pack_xdsl_hubic_loading_error": "Oups ! Nous n'avons pas pu lister vos services HubiC.",
   "promotionCode": "Bon de réduction",
@@ -47,10 +47,10 @@
   "used": "utilisé(s)",
   "used_one": "<strong>{{count}}</strong> utilisé",
   "used_many": "<strong>{{count}}</strong> utilisés",
-  "coming_soon": "Próximamente disponible",
+  "coming_soon": "Coming soon",
   "hostedEmail": "Hosted Emails",
   "emailPro": "xDSL E-mail",
-  "my_service": "Mis servicios",
+  "my_service": "My services",
   "manage_my_pack": "Gérer mon pack",
   "account": "compte",
   "size": "taille",
@@ -60,16 +60,16 @@
   "pack_xdsl_ok": "Pack actif",
   "pack_xdsl_expired": "Pack expiré",
   "pack_xdsl_inCreation": "Pack en cours de création",
-  "pack_xdsl_offer_name": "Nombre del producto",
+  "pack_xdsl_offer_name": "Service name",
   "pack_xdsl_pack_id": "Identifiant du pack",
-  "pack_xdsl_created_date": "Fecha de creación",
+  "pack_xdsl_created_date": "Creation date",
   "pack_xdsl_engaged_date": "Engagé jusqu'au",
   "pack_xdsl_not_engaged": "Vous n'êtes plus engagé",
   "pack_xdsl_resiliation_date": "Date de résiliation",
   "pack_xdsl_pack_link_address": "Adresse de raccordement",
   "still_pending": "en cours de création",
-  "account_name": "Email",
-  "all_fields_are_required": "Todos los campos son obligatorios.",
+  "account_name": "Email address",
+  "all_fields_are_required": "All fields are mandatory ",
   "required_fields_are_tag_with_a_star": " Les champs obligatoires sont signalés par un astérisque",
   "in_validation": "Votre demande est en cours",
   "title_min_size": "Le nom du compte doit comporter au minimum 5 caractères",
@@ -84,6 +84,6 @@
   "pack_promotion_code_condition": "Je comprends qu'en générant mon bon d'achat de <strong>{{text}}</strong> HT, je me réengage sur <b>{{engagement}} mois</b> pour le <b>pack complet</b>. Cette opération est <b>irréversible</b>.",
   "pack_promotion_code_confirm": "Je me réengage",
   "pack_rename_error": "Impossible de renommer votre pack {{serviceName}}.",
-  "pack_offer_migration": "Changer d'offre",
+  "pack_offer_migration": "Upgrade",
   "pack_xdsl_oops_an_error_is_occured": "Oups ! Nous n'avons pas pu récupérer les informations relatives à votre pack."
 }
