+++ packages/manager/apps/telecom/src/app/telecom/pack/dashboard/translations/Messages_es_ES.json
