+++ packages/manager/apps/telecom/src/app/telecom/pack/dashboard/translations/Messages_es_US.json
@@ -10,14 +10,14 @@
   "pack_xdsl_avalaibleSlot_0": "Vous avez consommé tout vos {{packType}} sur ce pack",
   "pack_xdsl_avalaibleSlot_1": "{{avalaible}} de disponible",
   "pack_xdsl_avalaibleSlot_n": "{{avalaible}} de disponibles",
-  "pack_xdsl_access_services_0": "Brak dostępu",
-  "pack_xdsl_access_description": "Opis",
-  "pack_xdsl_access_type": "Typ",
-  "pack_xdsl_access_accessName": "Identyfikator",
-  "pack_xdsl_voipLine_domain": "Domena",
-  "pack_xdsl_voipLine_billingAccount": "Konto",
-  "pack_xdsl_hubic_domain": "Domena",
-  "pack_xdsl_hubic_size": "Rozmiar",
+  "pack_xdsl_access_services_0": "No hay accesos",
+  "pack_xdsl_access_description": "Descripción",
+  "pack_xdsl_access_type": "Tipo",
+  "pack_xdsl_access_accessName": "Identifiant",
+  "pack_xdsl_voipLine_domain": "Dominio",
+  "pack_xdsl_voipLine_billingAccount": "Compte",
+  "pack_xdsl_hubic_domain": "Dominio",
+  "pack_xdsl_hubic_size": "Taille",
   "pack_xdsl_hubic_voucher": "(Code : {{voucher}})",
   "pack_xdsl_hubic_is_used": "Déja utilisé",
   "pack_xdsl_hubic_used_voucher_pending_details": "Récupération des données...",
@@ -26,11 +26,11 @@
   "pack_xdsl_hubic_domain_details_error": "Une erreur est survenue lors de la récupération du détail des domaines",
   "xdslAccess": "Accès Internet",
   "xdslAccess_default_sidebar_title": "Mes accès",
-  "voipLine": "Telefonia",
-  "domain": "Domena",
+  "voipLine": "Téléphonie",
+  "domain": "Nom de domaine",
   "voipEcoFax": "EcoFax Pro",
   "hubic": "HubiC",
-  "task": "Operacje",
+  "task": "Opérations",
   "pack_xdsl_hubic_activate": "Profiter de mon offre hubiC",
   "pack_xdsl_hubic_loading_error": "Oups ! Nous n'avons pas pu lister vos services HubiC.",
   "promotionCode": "Bon de réduction",
@@ -47,10 +47,10 @@
   "used": "utilisé(s)",
   "used_one": "<strong>{{count}}</strong> utilisé",
   "used_many": "<strong>{{count}}</strong> utilisés",
-  "coming_soon": "Dostępne wkrótce",
+  "coming_soon": "Próximamente disponible",
   "hostedEmail": "Hosted Emails",
   "emailPro": "xDSL E-mail",
-  "my_service": "Moje usługi",
+  "my_service": "Mis servicios",
   "manage_my_pack": "Gérer mon pack",
   "account": "compte",
   "size": "taille",
@@ -60,16 +60,16 @@
   "pack_xdsl_ok": "Pack actif",
   "pack_xdsl_expired": "Pack expiré",
   "pack_xdsl_inCreation": "Pack en cours de création",
-  "pack_xdsl_offer_name": "Nazwa oferty",
+  "pack_xdsl_offer_name": "Nombre del producto",
   "pack_xdsl_pack_id": "Identifiant du pack",
-  "pack_xdsl_created_date": "Data utworzenia",
+  "pack_xdsl_created_date": "Fecha de creación",
   "pack_xdsl_engaged_date": "Engagé jusqu'au",
   "pack_xdsl_not_engaged": "Vous n'êtes plus engagé",
   "pack_xdsl_resiliation_date": "Date de résiliation",
   "pack_xdsl_pack_link_address": "Adresse de raccordement",
   "still_pending": "en cours de création",
-  "account_name": "E-mail",
-  "all_fields_are_required": "Wszystkie pola są obowiązkowe",
+  "account_name": "Email",
+  "all_fields_are_required": "Todos los campos son obligatorios.",
   "required_fields_are_tag_with_a_star": " Les champs obligatoires sont signalés par un astérisque",
   "in_validation": "Votre demande est en cours",
   "title_min_size": "Le nom du compte doit comporter au minimum 5 caractères",
@@ -83,7 +83,7 @@
   "pack_promotion_code_modal_title": "Profiter du bon de réduction",
   "pack_promotion_code_condition": "Je comprends qu'en générant mon bon d'achat de <strong>{{text}}</strong> HT, je me réengage sur <b>{{engagement}} mois</b> pour le <b>pack complet</b>. Cette opération est <b>irréversible</b>.",
   "pack_promotion_code_confirm": "Je me réengage",
-  "pack_rename_error": "Nie można zmienić nazwy pakietu {{serviceName}}.",
-  "pack_offer_migration": "Zmień ofertę",
+  "pack_rename_error": "Impossible de renommer votre pack {{serviceName}}.",
+  "pack_offer_migration": "Changer d'offre",
   "pack_xdsl_oops_an_error_is_occured": "Oups ! Nous n'avons pas pu récupérer les informations relatives à votre pack."
 }
