+++ packages/manager/apps/telecom/src/app/telecom/pack/dashboard/translations/Messages_fr_CA.json
