+++ packages/manager/apps/telecom/src/app/telecom/pack/dashboard/translations/Messages_it_IT.json
