+++ packages/manager/apps/telecom/src/app/telecom/pack/dashboard/translations/Messages_pt_PT.json
