--- packages/manager/apps/telecom/src/app/telecom/pack/internet-access/internet-access.component.js
@@ -1,5 +1,0 @@
-import template from './internet-access.html';
-
-export default {
-  template,
-};
