--- packages/manager/apps/telecom/src/app/telecom/pack/internet-access/internet-access.html
@@ -1,5 +1,0 @@
-<oui-header
-    data-heading="{{ :: 'pack_internet_access_title' | translate }}"
-></oui-header>
-
-<div data-ui-view></div>
