--- packages/manager/apps/telecom/src/app/telecom/pack/internet-access/internet-access.routing.js
@@ -1,11 +1,0 @@
-export default /* @ngInject */ ($stateProvider) => {
-  $stateProvider.state('telecom.packs.internet-access', {
-    url: '',
-    abstract: true,
-    views: {
-      'packView@telecom.packs': {
-        component: 'telecomPackInternetAccess',
-      },
-    },
-  });
-};
