--- packages/manager/apps/telecom/src/app/telecom/pack/internet-access/packs/index.js
@@ -1,14 +1,0 @@
-import angular from 'angular';
-
-import component from './packs.component';
-import routing from './packs.routing';
-
-const moduleName = 'ovhManagerTelecomPackInternetAccessPacks';
-
-angular
-  .module(moduleName, [])
-  .component('telecomPackInternetAccessPacks', component)
-  .config(routing)
-  .run(/* @ngTranslationsInject:json ./translations */);
-
-export default moduleName;
