--- packages/manager/apps/telecom/src/app/telecom/pack/internet-access/packs/translations/Messages_fr_FR.json
@@ -1,6 +1,0 @@
-{
-  "packs_pack_name_label": "Pack",
-  "packs_description_label": "Description",
-  "packs_num_service_label": "Accès Internet",
-  "packs_view_label": "Détail du pack"
-}
