--- packages/manager/apps/telecom/src/app/telecom/pack/internet-access/translations/Messages_fr_FR.json
@@ -1,3 +1,0 @@
-{
-  "pack_internet_access_title": "Accès Internet"
-}
