+++ packages/manager/apps/telecom/src/app/telecom/pack/packs.component.js
