+++ packages/manager/apps/telecom/src/app/telecom/pack/packs.controller.js
