+++ packages/manager/apps/telecom/src/app/telecom/pack/packs.html
@@ -1,3 +1,6 @@
+<oui-header
+    data-heading="{{ :: 'pack_internet_access_title' | translate }}"
+></oui-header>
 <div class="container-fluid my-5">
     <oui-datagrid
         id="dg-telephony-packs"
