+++ packages/manager/apps/telecom/src/app/telecom/pack/translations/Messages_fr_FR.json
@@ -1,89 +1,7 @@
 {
-  "pack_xdsl_page_title": "Pack \"{{name}}\"",
-  "pack_xdsl_service_list": "Gestion de votre pack {{packName}}",
-  "pack_xdsl_service_more_one": "{{count}} supplémentaire",
-  "pack_xdsl_service_more_many": "{{count}} supplémentaires",
-  "pack_xdsl_access": "Mes accès",
-  "pack_xdsl_no_service_available": "Aucun service disponible",
-  "pack_xdsl_resiliation_success_notification": "Votre demande de résiliation est prise en compte. Si vous changez d'avis, sachez que vous pouvez toujours annuler votre demande.",
-  "pack_xdsl_resiliation_cancel_success_notification": "Votre demande de résiliation est annulée. Quel plaisir de vous garder parmi nous.",
-  "pack_xdsl_avalaibleSlot_0": "Vous avez consommé tout vos {{packType}} sur ce pack",
-  "pack_xdsl_avalaibleSlot_1": "{{avalaible}} de disponible",
-  "pack_xdsl_avalaibleSlot_n": "{{avalaible}} de disponibles",
-  "pack_xdsl_access_services_0": "Aucun accès",
-  "pack_xdsl_access_description": "Description",
-  "pack_xdsl_access_type": "Type",
-  "pack_xdsl_access_accessName": "Identifiant",
-  "pack_xdsl_voipLine_domain": "Domaine",
-  "pack_xdsl_voipLine_billingAccount": "Compte",
-  "pack_xdsl_hubic_domain": "Domaine",
-  "pack_xdsl_hubic_size": "Taille",
-  "pack_xdsl_hubic_voucher": "(Code : {{voucher}})",
-  "pack_xdsl_hubic_is_used": "Déja utilisé",
-  "pack_xdsl_hubic_used_voucher_pending_details": "Récupération des données...",
-  "pack_xdsl_hubic_used_email": "Adresse email : {{email}}",
-  "pack_xdsl_hubic_used_email_unavailable": "Donnée indisponible",
-  "pack_xdsl_hubic_domain_details_error": "Une erreur est survenue lors de la récupération du détail des domaines",
-  "xdslAccess": "Accès Internet",
-  "xdslAccess_default_sidebar_title": "Mes accès",
-  "voipLine": "Téléphonie",
-  "domain": "Nom de domaine",
-  "voipEcoFax": "EcoFax Pro",
-  "hubic": "HubiC",
-  "task": "Opérations",
-  "pack_xdsl_hubic_activate": "Profiter de mon offre hubiC",
-  "pack_xdsl_hubic_loading_error": "Oups ! Nous n'avons pas pu lister vos services HubiC.",
-  "promotionCode": "Bon de réduction",
-  "exchangeAccount": "Exchange (Compte)",
-  "exchangeLite": "Exchange (Lite)",
-  "exchangeOrganization": "Exchange (Organisation)",
-  "siteBuilderFull": "Site Builder (Complet)",
-  "siteBuilderStart": "Site Builder",
-  "voipAlias": "Alias",
-  "informations": "Informations Générales",
-  "exchangeIndividual": "Exchange (Individuel)",
-  "add": "ajouter",
-  "no_item": "aucun élément",
-  "used": "utilisé(s)",
-  "used_one": "<strong>{{count}}</strong> utilisé",
-  "used_many": "<strong>{{count}}</strong> utilisés",
-  "coming_soon": "Bientôt disponible",
-  "hostedEmail": "Hosted Emails",
-  "emailPro": "xDSL E-mail",
-  "my_service": "Mes services",
-  "manage_my_pack": "Gérer mon pack",
-  "account": "compte",
-  "size": "taille",
-  "move": "Déménager",
-  "pack_xdsl_information_associated_line": "Ligne associée",
-  "pack_xdsl_status": "État du pack",
-  "pack_xdsl_ok": "Pack actif",
-  "pack_xdsl_expired": "Pack expiré",
-  "pack_xdsl_inCreation": "Pack en cours de création",
-  "pack_xdsl_offer_name": "Nom de l'offre",
-  "pack_xdsl_pack_id": "Identifiant du pack",
-  "pack_xdsl_created_date": "Date de création",
-  "pack_xdsl_engaged_date": "Engagé jusqu'au",
-  "pack_xdsl_not_engaged": "Vous n'êtes plus engagé",
-  "pack_xdsl_resiliation_date": "Date de résiliation",
-  "pack_xdsl_pack_link_address": "Adresse de raccordement",
-  "still_pending": "en cours de création",
-  "account_name": "Email",
-  "all_fields_are_required": "Tous les champs sont obligatoires",
-  "required_fields_are_tag_with_a_star": " Les champs obligatoires sont signalés par un astérisque",
-  "in_validation": "Votre demande est en cours",
-  "title_min_size": "Le nom du compte doit comporter au minimum 5 caractères",
-  "you_dont_have_any_service": "Vous n'avez pas encore de services actifs.",
-  "create_a_new_service_with_the_adding_button": "Activez un nouveau service grâce au lien \"Activer un service\".",
-  "create_a_new_hubic_with_the_adding_button": "Activez un nouveau service grâce au lien \"Profiter de mon offre hubiC\".",
-  "pack_xdsl_loading": "Le chargement de votre pack est en cours ...",
-  "pack_xdsl_this_pack_seems_to_be_unavailable": "Il semblerait que ce pack n'existe pas.",
-  "pack_promotion_code_proposal": "OVH vous offre un bon d'achat d'une valeur de <strong>{{text}}</strong> HT, valable sur l'ensemble de nos produits (hors forfait ADSL-VDSL/SDSL, hors offre hubiC, hors achat de matériel et caution de téléphone, hors crédit du compte client OVH et achat de points prépayés).",
-  "pack_promotion_code_question": "J'en profite",
-  "pack_promotion_code_modal_title": "Profiter du bon de réduction",
-  "pack_promotion_code_condition": "Je comprends qu'en générant mon bon d'achat de <strong>{{text}}</strong> HT, je me réengage sur <b>{{engagement}} mois</b> pour le <b>pack complet</b>. Cette opération est <b>irréversible</b>.",
-  "pack_promotion_code_confirm": "Je me réengage",
-  "pack_rename_error": "Impossible de renommer votre pack {{serviceName}}.",
-  "pack_offer_migration": "Changer d'offre",
-  "pack_xdsl_oops_an_error_is_occured": "Oups ! Nous n'avons pas pu récupérer les informations relatives à votre pack."
+  "packs_pack_name_label": "Pack",
+  "packs_description_label": "Description",
+  "packs_num_service_label": "Accès Internet",
+  "packs_view_label": "Détail du pack",
+  "pack_internet_access_title": "Accès Internet"
 }
