+++ packages/manager/apps/telecom/src/components/sidebar/sidebar.config.js
@@ -4,7 +4,7 @@ import { FeatureAvailability } from '@ovh-ux/ng-ovh-telecom-universe-components'
 import { FREEFAX_AVAILABILITY } from '@ovh-ux/manager-freefax';
 import { OTB_AVAILABILITY } from '@ovh-ux/manager-overthebox';
 import { ALIAS_AVAILABILITY } from '../../app/telecom/orders/alias/orders-alias.constants';
-import { PACK_AVAILABILITY } from '../../app/telecom/pack/pack.constant';
+import { PACK_AVAILABILITY } from '../../app/telecom/pack/dashboard/pack.constant';
 import { TELEPHONY_AVAILABILITY } from '../../app/telecom/telephony/telecom-telephony.constant';
 
 import { URLS } from './sidebar.constants';
